+++ lib/protocol/content.rb
@@ -6,7 +6,6 @@
 require_relative "content/version"
 require_relative "content/error"
 require_relative "content/parser"
-require_relative "content/representation"
 
 module Protocol
 	# @namespace