+++ test/protocol/content/parser.rb
@@ -33,4 +33,51 @@ describe Protocol::Content::Parser do
 			parser.register("text/plain"){}
 		end.to raise_exception(FrozenError)
 	end
+	
+	it "parses content using a compatible handler" do
+		media_type = nil
+		input = Object.new
+		parser = subject.build do |parser|
+			parser.register("text/*") do |candidate, parsed_media_type|
+				media_type = parsed_media_type
+				candidate
+			end
+		end
+		
+		expect(parser.parse("text/plain; charset=utf-8", input)).to be_equal(input)
+		expect(media_type.name).to be == "text/plain"
+		expect(media_type.parameters).to be == {"charset" => "utf-8"}
+	end
+	
+	it "forwards a block to the content handler" do
+		parser = subject.build do |parser|
+			parser.register("text/plain") do |input, _media_type, &block|
+				block.call(input)
+			end
+		end
+		
+		value = parser.parse("text/plain", "hello") do |input|
+			input.upcase
+		end
+		
+		expect(value).to be == "HELLO"
+	end
+	
+	it "rejects unsupported media types" do
+		parser = subject.new
+		
+		expect do
+			parser.parse("text/plain", Object.new)
+		end.to raise_exception(Protocol::Content::UnsupportedMediaTypeError) do |error|
+			expect(error.media_type.name).to be == "text/plain"
+		end
+	end
+	
+	it "rejects a missing media type" do
+		parser = subject.new
+		
+		expect do
+			parser.parse(nil, Object.new)
+		end.to raise_exception(Protocol::Content::UnsupportedMediaTypeError, message: be == "Missing media type!")
+	end
 end